Rename `models.data` to `models.db` which is more specific
