+++ checkmate/models/db/custom_rule.py
