--- checkmate/models/domain/__init__.py
@@ -1,1 +1,0 @@
-"""Logical domain model objects."""
